Move big_cty into db loaders
